requirements/install.txt
Django==1.6.2
South==0.8.4
argparse==1.2.1
django-oauth2-provider==0.2.6
django-treebeard==2.0rc1
djangorestframework==2.3.12
shortuuid==0.4
six==1.5.2
wsgiref==0.1.2
